Improve recipes documentation * Rename input-1 to custom-text-input * Add to each recipe readme a usage example and props description
