+++ cookbook-react-native/recipes/buttons/custom-highlight-button/cookbook.json
@@ -1,0 +1,10 @@
+{
+  "thumbnail": {
+    "type": "img",
+    "url": "https://raw.githubusercontent.com/Wolox/frontend-cookbook/master/cookbook-react-native/recipes/buttons/custom-highlight-button/recipe.gif"
+  },
+  "detail": {
+    "type": "img",
+    "url": "https://raw.githubusercontent.com/Wolox/frontend-cookbook/master/cookbook-react-native/recipes/buttons/custom-highlight-button/recipe.gif"
+  }
+}
