+++ cookbook-react-native/recipes/buttons/custom-highlight-button/utils.ts
