--- cookbook-react-native/recipes/buttons/custom-hightlight-button/cookbook.json
@@ -1,10 +1,0 @@
-{
-  "thumbnail": {
-    "type": "img",
-    "url": "https://raw.githubusercontent.com/Wolox/frontend-cookbook/master/cookbook-react-native/recipes/buttons/custom-hightlight-button/recipe.gif"
-  },
-  "detail": {
-    "type": "img",
-    "url": "https://raw.githubusercontent.com/Wolox/frontend-cookbook/master/cookbook-react-native/recipes/buttons/custom-hightlight-button/recipe.gif"
-  }
-}
