+++ cookbook-react-native/recipes/buttons/custom-opacity-button/constants.ts
@@ -4,17 +4,17 @@ import { CustomTextVariants } from '@textsRecipes/custom-text/constants';
 export const VARIANTS = ['borderless', 'radial', 'black', 'green', 'white', 'gray'];
 
 interface CustomButtonVariants {
+  black?: boolean;
   borderless?: boolean;
   radial?: boolean;
-  black?: boolean;
 }
 
 export interface CustomButtonProps extends CustomTextVariants, CustomButtonVariants {
-  onPress: TouchableOpacityProps['onPress'];
   activeOpacity?: number;
   disabled?: boolean;
   icon?: number;
   iconStyle?: ImageStyle;
+  onPress: TouchableOpacityProps['onPress'];
   style?: ViewStyle;
   textStyle?: TextStyle;
   title?: string;
