+++ cookbook-react-native/recipes/inputs/custom-text-input/assets/ic_visibility_off.png
