+++ cookbook-react-native/recipes/inputs/custom-text-input/components/InputLabel/index.tsx
