+++ cookbook-react-native/recipes/inputs/custom-text-input/components/ShowPassword/styles.ts
