+++ cookbook-react-native/recipes/inputs/custom-text-input/cookbook.json
@@ -1,0 +1,10 @@
+{
+  "thumbnail": {
+    "type": "img",
+    "url": "https://raw.githubusercontent.com/Wolox/frontend-cookbook/master/cookbook-react-native/recipes/inputs/custom-text-input/recipe.gif"
+  },
+  "detail": {
+    "type": "img",
+    "url": "https://raw.githubusercontent.com/Wolox/frontend-cookbook/master/cookbook-react-native/recipes/inputs/custom-text-input/recipe.gif"
+  }
+}
