+++ cookbook-react-native/recipes/inputs/custom-text-input/interfaces.ts
@@ -1,26 +1,23 @@
-import { TextProps, TextInputProps, TextStyle, ViewProps } from 'react-native';
+import { TextProps, TextInputProps, TextStyle, ViewStyle } from 'react-native';
 
 export interface CustomTextInputProps extends TextInputProps, TextProps {
   animated?: boolean;
-  autoCompleteType?: TextInputProps['autoCompleteType'];
   disabled?: boolean;
-  error?: boolean | string;
-  errorContainerStyle?: ViewProps['style'];
+  error?: string;
+  errorContainerStyle?: ViewStyle;
   errorStyle?: TextStyle;
-  inputContainerStyle?: ViewProps['style'];
+  inputContainerStyle?: ViewStyle;
   inputTextStyles?: TextStyle;
   isFocused: boolean;
   label?: string;
   labelStyle?: TextStyle;
   multiline?: boolean;
-  onBlur?: TextInputProps['onBlur'];
   onChange?(e: any): any;
-  onFocus?: TextInputProps['onFocus'];
   placeholder?: string;
   placeholderColor?: string;
   secureTextEntry?: boolean;
   showError?: boolean;
   showEye?: boolean;
-  style?: ViewProps['style'];
+  style?: ViewStyle;
   value?: string;
 }
