+++ cookbook-react-native/recipes/inputs/custom-text-input/recipe.gif
