+++ cookbook-react-native/recipes/inputs/custom-text-input/styles.ts
