--- cookbook-react-native/recipes/inputs/input-1/cookbook.json
@@ -1,10 +1,0 @@
-{
-  "thumbnail": {
-    "type": "img",
-    "url": "https://raw.githubusercontent.com/Wolox/frontend-cookbook/master/cookbook-react-native/recipes/inputs/input-1/recipe.gif"
-  },
-  "detail": {
-    "type": "img",
-    "url": "https://raw.githubusercontent.com/Wolox/frontend-cookbook/master/cookbook-react-native/recipes/inputs/input-1/recipe.gif"
-  }
-}
