+++ cookbook-react-native/recipes/texts/custom-text/constants.ts
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
-import { TextProps } from 'react-native';
+import { TextProps, TextStyle } from 'react-native';
 
 /*
  ** TODO: You can add styles to Base like Family Font to be the Text styles base!
@@ -53,5 +53,5 @@ export interface CustomTextVariants {
 export interface CustomTextProps extends CustomTextVariants {
   children: ReactNode;
   textProps?: TextProps;
-  style?: any;
+  style?: TextStyle;
 }
